fine

+++ scripts/02.sh
@@ -5,8 +5,7 @@ then
 	echo "Enter something please"
 	exit 1
 else
-	if [ "$(getent passwd $1)" ]
-	then
+	if [ "$(getent passwd $1)" ] && [ $1 != "nobody" ]; then
 		uid= `id -u $1`
 		pkill -U $1
 		userdel -r -f "$1"
